+++ api/index.js
@@ -5,7 +5,7 @@ const { DB_PORT } = process.env;
 // require('dotenv').config();
 // const { PGPORT } = process.env;
 
-conn.sync({ force: true }).then(() => {
+conn.sync({ force: false }).then(() => {
   server.listen(DB_PORT, () => {
     console.log(`%s listening at ${DB_PORT}`); // eslint-disable-line no-console
   });
